Added check of surrounding whitespace
This is meant to provide user with the additional option of checking
the translation based on the whitespace surrounding the actual text.
The text in the translation is considered good only if it ends and begins
with same whitespace characters.

The Binding spacial characters in Unicode are considered as well.

[ChangeLog][Linguist] Added capability of checking that the whitespace
surrounding the translations' text matches that of the source strings.

Task-number: QTBUG-16368
Change-Id: Ieb92b33809e0eccb95c8dfee60a0f4f9b3b2c588
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>