Use the size hint of the pushbutton when determining the minimum height
Since the pushbutton has not been shown at the time the height of it is
requested then it has a default which is bigger than it will ultimately
be. Therefore to ensure the layout is setup correctly we depend on the
sizeHint() instead to give the right height for the button.

Task-number: QTBUG-52020
Change-Id: I569bfbc59671e682c83e98d38199aca69c6b768f
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>