qdoc: Reduce total clang parse time
Including the module's private headers in the PCH
database can reduce the qdoc run time for the module
by a few minutes. Apparently, including the private
headers significantly reduces the number of header
files that must be re-parsed.

This change adds the module's private headers to the
pre-compiled header database. When it finds the module's
module header, it copies that module header to the
temporary directory where it will construct the PCH.
Then it finds the module's private header subdirectory,
constructs the path to each header in that directory,
and appends it to the temporary module header. Then
it passes this augmented module header to clang to
construct the PCH.

Change-Id: Ie67485c7070ef7487345db90a8b27c64f5caa0f2
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>