Doc: Assistant: Update internal .qch help file
... and introduce a doc/internal project configuration,
that can be used for manually updating the assistant.qch file.

This way, it's easier to maintain changes to the template
(style, copyright info) and the doc content.

Separate the 'Qt Assistant Quick Guide' into its own .qdoc
file, so it can be used by both the Qt Assistant Manual and
the internal documentation set. Use \if commands to remove
navigation from the internal doc, as it cannot have links
to external docs (that may not be registered yet).

Task-number: QTBUG-51709
Change-Id: Id832f7911e9a56958c8b9e34409bd96efc19554f
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>