lupdate: Fix parsing of class template specializations
4ea581c95ae9704b250591c5d4d9b0517f5b62f2 introduced a bug where
class template specialization would lead to contexts not being properly
cleaned up on '}'. The occurrence of class template specializations
lead to a massive slowdown in parsing following includes because they
were treated as direct (in-namespace) and parsed more often than
normally needed.

Unfortunately the bug was not discovered early by the auto tests,
but eventually the slowdown appeared in CI when merging to dev.
This change adds a test for class template specializations. It would
discover the cause (broken context) which triggers the slowdown but
doesn't depend on includes.

Task-number: QTBUG-53644
Change-Id: I94cce58702b53dfb5e0f9c92585222dd3f938593
Reviewed-by: Liang Qi <liang.qi@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>