Ensure that Nbsp characters are preserved by escaping them
Characters that are considered spaces, but aren't just ' ' should always
be escaped so that the exact type of space that they are is preserved.
Since QTextDocument::toPlainText() will convert QChar::Nbsp to ' ', we
need to get the plain text from the handler directly and do our own
conversion where applicable to stop it from replacing the Nbsp characters.

Task-number: QTBUG-35652
Change-Id: I373edca92253ca5f978ba656c8ebbcaf77965e61
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>