macdeployqt: add library search path
In earlier versions of OSX, the folder '/usr/lib/' is protected by
default and user libraries cannot be copied to that folder.
Alternatively, providing an option for alternative library search path
would avoid missing libraries during the deploy.

The addition of the option '-libpath=<path>' to macdeployqt adds a
search path for the dependencies. If a dependency was not found in the
provided path, the default folder for searching is used.

[ChangeLog][macdeployqt] Added option '-libpath' to search libraries in
a custom path.

Change-Id: I3afee6aaa3bc25521adecb91ec59d104e13834ed
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>