Less randomization in output of qdoc.
There is a lot of random order going on in the output. It makes doing
two runs of documentation hard to compare, because things comes in different order.

It should be easier to debug the output now that lines and elements are written
in a sorted order.

There are still some way towards reproducible qdoc.

Change-Id: Ia3f64d5463cc812e13c25ca84c47906c75093f26
Reviewed-by: David Faure <david.faure@kdab.com>