Skip deployment of dependencies with lower version number
Depending on the setup there might be a newer version of a dependency on
the target device compared to the local setup. For instance one can have
a preview build installed, but the development environment runs on a
stable version.

In that case deployment of the C-Runtime is rejected as a newer version
is available. Instead of returning an error, report a warning and
continue, relying on the fact that the newer version stays binary
compatible. Unfortunately, there is no "proper" error code for this, so
we need to use the value itself.

Task-number: QTBUG-54251
Change-Id: I83b66653c776665a4aae6ab138d8d292def426f5
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>