winrt: arm: copy ucrtbased.dll in debug mode
For Windows 10 / MSVC2015 winrtrunner fails to run an app in debug mode
on an arm device as symbols were not found (neither in the system or c-
runtime dependency). Those symbols are inside ucrtbased.dll. Visual
Studio 2015 silently adds this library to the package resulting in a
different content compared to windeployqt.

Hence, copy ucrtbased.dll manually from windeployqt as well, which will
later on be packaged by winrtrunner.

Task-number: QTBUG-52969
Change-Id: If43a814b800d0756d111557a042aa23a22d7589f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>