winrtrunner: Fix deployment of QtQuick applications when using appx
For remote devices winrtrunner needs to create an appx package for
deployment. While creating that package a file extension list has been
used to identify potential files by Qt which need to be deployed.

This was broken for Qt Quick Controls applications, but also for
projects using for instance jpeg image files.

Instead, switch to an exclusion list of file extensions we can be sure
should not be part of the package. In the worst case the package is too
big when deployed with winrtrunner, but will not cause runtime errors
anymore.

Change-Id: Ief6d4aaf9bc0933b32734feacbef170dcc8ab945
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>