make include(,<into>) implementation a bit less convoluted
this goes at the cost of <into> without the subsequent dot not being
filtered before the result insertion any more. that's not expected to
be a problem, as it's not particularly useful to iterate the target
namespace without the dot in the first place.

Change-Id: Ib3497a60613b4c8c26676fe76b731239e427b8a7
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from qtbase/3277143f9609a0c257dd118c7fa4900b2fb36de3)
Reviewed-by: Jake Petroules <jake.petroules@qt.io>