qdoc: Use group information for page navigation (breadcrumbs)
Improve the contents of the navigation bar for pages generated with
the \page command: If the \page is a member of a single group (uses
a single \ingroup command), the related \group page exists and has a
title, then inject a link to the group page into the navigation bar.

Pages can be members of multiple groups; if so, we do not use the
information for navigation as there's no mechanism to determine the
'primary' group.

Change-Id: I7e7e3167b3a90a94fd7d0a14f412046310ad9414
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>