qdoc: Find the system include path if not provided
clang needs the path to the system include files. There is
an option to pass system include paths to qdoc on the command
line, but currently there is no way to get the information
from the build system.

With this change, clangqdoc asks if the system include paths
have been provided. If not, it uses the path to <locale.h> as
the system include path, which it passes to clang using the
-isystem option.

Thanks to Edward Welbourne for this solution.

Change-Id: I2716505126d7780fc616b8ffd7cd2865576bae16
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>