Fix a leak and a crash in designer
This change downports the
56ac0c12ab5e7269e16bcc1248e4241302d91d09
change to the 5.7.1 branch in order to
have a chance for releasing Qt Creator 4.2
with a fix. The crash doesn't appear in
the standalone Designer, it only appears
in the Designer plugin of Qt Creator during reload.
There is no other way to fix that crash
outside of Qt.

Task-number: QTCREATORBUG-17150
Change-Id: Ide521677a1526d3de862d0d8e283531a67b2a7a8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 56ac0c12ab5e7269e16bcc1248e4241302d91d09)