pixeltool: Check whether initial geometry is applicable
When changing primary screens of multi-monitor-setups on Windows,
out-of-screen positions might be saved which make it impossible to bring
the tool back and also cause crashes when no screen is found for grabbbing.

Check the saved position when reading the settings and check if a screen
can be obtained for grabbing.

Change-Id: I69ec677e144306290a4ee7a3e84f87dd600ff086
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>