qdoc: Handle class decls outside the semantic parent
We have the class declaration for QMetaObject::Connection
in qobjectdefs.h, which is outside the declaration for
QMetaObject. clangqdoc wasn't handling this correctly. It
created the class node for Connection with the wrong parent.

This update ensures that the class node for Connection is given
the class node of QMetaObject as its parent. If the semantic and
lexical parent cursors of the current cursor are not the same,
find the Aggregate node for the semantic parent cursor and use
that node as the parent when the new class node is created.

Change-Id: I40f73dad9879e39452f83bb7c0f514a7ef319208
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>