Fix a crash on Designer exit
After the destructor of FormWindow itself has finished,
it starts destructing its fields. One of them is
an undo stack. Unfortunately, the destructor of
QUndoStack emits signals. One of its signals is
being received in QDesignerFormWindow::updateChanged().
The QDesignerFormWindow keeps a QPointer<FormWindow> m_editor
which we are currently destructing. The destructor of
FormWindow finished, but the destructor of its base QObject
didn't run yet, so all QPointer instances pointing to it
are not nulled yet. Inside the updateChanged() we ensure that m_editor is
still alive, however this check is wrong in Qt5 times.
The QPointer will be nulled when the destructor of
base QObject finishes, so we read that m_editor is valid,
while it is in the middle of destruction. Then we call
some method on it which leads to the crash.

The current fix disconnects the undo stack's signals
prior to destuction. It also removes the undo stack
from the undo group.

Task-number: QTBUG-57201
Change-Id: I753b5c8cae2b630d7cdec13b0dfe325d16726e82
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>