qtattributionscanner: Support optional QtParts property
We'd like to differentiate between code that ends up in the
Qt libraries or plugins, or code that only is added to examples,
tests, or (host) tools. This can now be specified with the
JSON property 'QtParts'.

If 'QtParts' does not contain 'libs', the page will not show up in the
list of attributions generated by module with e.g.

  \generatelist{groupsbymodule attributions-qtcore}

nor in the global list generated by

  \generatelist annotatedattributions

Instead, third party code in examples or tests can now be documented
separately using e.g.

  \generatelist{groupsbymodule attributions-examples}

and

  \annotatedlist attributions-examples

Change-Id: Iba72f76b63e4435af54a61b7b4a2cf4617746ba6
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>