Be more explicit about linguist's module dependencies
The conditional addition of printsupport made it seem like linguist
was intended to compile also without printsupport. However,
QPrintDialog is extensively used, with no #ifdef guards whatsoever.
Therefore, printsupport is apparently not optional. It doesn't make a
lot of sense to build desktop tools with reduced Qt configurations.
Doing so would clutter the code with #ifdefs for very little benefit.

Change-Id: I815a0cfa90da6c302e894e0c10c18184128123fa
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>