qdoc: A different way to fix the "signed" problem
The old qdoc parser is still used for parsing \fn
commands because clang can't be used to parse code
fragments. This causes a problem when the "signed"
reserved word is used because clang doesn't keep
it if it doesn't add anything to the type, while
the old qdoc parser always appends it to the type
under construction.

The fix here is to let the old qdoc parser hold
the "signed" reserved word aside until it knows
what the next lexeme is. Then it either appends
the held "signed" or discards it, depending on
what the next lexeme is. e.g., if the next word
is "int" discard "signed" but if the next word
is "char" append the held "signed" before
appending "char".

Change-Id: I67953d964cd09410126d64fa3305c669111e74ce
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>