qdoc: Don't report missing parameter name
See QTBUG-58277. No longer report missing parameter names when
processing the \fn command, because there are cases where they are not
required. This is an over-simplification of the problem, but it works
for now.

In the future, we may want to distinguish between cases where the
parameter type is std::nullptr_t, which means the actual parameter is
null, and the case where some type is provided without a name, which
indicates that the parameter is not used in this implementation of the
function.

Task-number: QTBUG-58281
Change-Id: I7d648341262040b9b9ea91f1e27e5905455f9ef2
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>