qdoc: Use documentation for implicit special functions
Allow implicitly declared special member functions to be
explicitly documented. For example, in class QFuture, the
destructor, copy constructor and assignment operatore are
implicitly declared, but they are explicitly documented
in qfuture.qdoc. clangqdoc no longer complains that it
can't find these functions in any header file. It creates
them for documentation purposes, and they are documented
as expected.

Where these function signatures appear on the class reference
page, " = default" is appended to the end to indicate that the
function is implicitly declared.

Change-Id: I7bcf16c31099439f424b438eb41589435d7254b6
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>