qdoc: Teach parser for \fn commands to parse array parameters
The old qdoc C++ parser is still used for parsing function
signatures in \fn commands. The parser could not handle the
case where a formal parameter is an array declaration. Now
it parses the formal parameter, so the syntax errors have
disappeared. But there is a separate problem involving the
matching of the declaration to the same declaration in the
header file, which is parsed by clang. However, this is a
separate issue not addressed by this fix.

Change-Id: I61be810b02ac50a6ee380664a41915cc85633c66
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>