qdoc: create the anonymous enum type
When clangqdoc encounters an unnamed enum type in a class,
qdoc reports an error that enum type "global" is not found
in any header file, and then the identifiers in the enum
type can't be properly documented. This change makes the
clang qdoc visitor name the unnamed enum type "anonymous"
and allows it to be documented like a named enum type.

Change-Id: I8b6dbc9bb78adc5f5c39a2a2d73c27ccb4543ceb
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>