winrtrunner: Fix waiting for local package (un-)installation
Debugging did not work on some machines as the waiting loop used in
qlocalappxengine was not implemented properly.
Instead of relying on an undocumented return value of GetResults we
register a callback that is triggered when the deployment operation
is finished and trigger an event inside that callback in order to
avoid a busy loop.

Task-number: QTCREATORBUG-17907
Change-Id: I0bd6535ef0e333ec27b9f339e0c786a63d3b9d07
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>