winrtrunner: Fix waiting for local package (un-)installation
Debugging did not work on some machines as the waiting loop used in
qlocalappxengine was not implemented properly.
Instead of relying on an undocumented return value of GetResults we
register a callback that is triggered when the deployment operation
is finished and trigger an event inside that callback in order to
avoid a busy loop.

Task-number: QTCREATORBUG-17907
Change-Id: I5d2211cfabb71592efa520cf53ad84facf79a4df
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
(cherry picked from commit 3947d229cb8ed8dde5ab72708a4a149ac6d83cae)