qdoc: Ignore declarations of anonymous structs
qdoc doesn't need to see declarations of anonymous structs.
When it does see one inside a class or struct that is named,
then the html generator tries to overwrite the html file for
the outer class or struct. That causes a few qdoc error
messages and it overwrites the html file for the class
reference page and the all-members page. This update tells
clang not to visit the declaration of the anonymous struct,
so the html generator doesn't see it as something to be
documented.

Change-Id: Ie8c732e650c39f78c1374523d72a7e0448a31ba3
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>