qtattributionsscanner: Allow to use it for one single file
This allows a saner use with qbs, where we can and should
actually handle the dependencies correctly.

In addition, this patch adds an optional '--basedir' argument
that is needed to correctly root relative paths that are
printed in the documentation.

Change-Id: I0abf769ad5c0a3ac15b6907f83b77d369eb78d1f
GPush-Base: 5c8d5742dc7b5e00a6aed447e8f288680b58e225
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>