Account for custom property sheet extensions when getting objectName
In order to account for custom property sheet extensions in a custom
widget plugin, it needs to be possible to obtain the objectName value
from a QVariant containing a QString. So to enable this, a check is
done to see if the variant is storing a PropertySheetStringValue. If
it is not then it is expected to be convertable to a string.

Task-number: QTBUG-62936
Change-Id: I8ce2053256a5b18851254c3ed1d6037d12023bf5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>