qdoc: QPrivateSignal was not handled correctly
Now that clang-qdoc uses clang to parse the \fn commands,
it must handle the QPrivateSignal flag correctly. It is not
allowed in the signature in the \fn command. This update
corrects a bug where the expected number or parameters was
being reduced by 1 because the flag appeared in the declaration
in the include file.

Change-Id: I1c974100b2017b3058040937a03500a75dc69d82
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>