qdoc: Revive the WebXML generator
Reintroduce the WebXML generator that was last seen in Qt 4.7 - this
generator produces XML output similar to .index files, with embedded
documentation content. To be used for generating documentation for the
PySide project.

Uses code from QDocIndexFiles to generate the elements. The generator
itself is a specialization of the HTML generator as a lot of the code
common to all generators is implemented in the HTML generator and not
the base class. This could be cleaned up later on.

To use the generator, set the 'outputformats' .qdocconf variable:

    outputformats = WebXML

In addition for PySide/Shiboken, \snippet commands must be
configured to generate <snippet> elements instead of hardcoded
<code> snippets:

    quotinginformation = true

Task-number: PYSIDE-363
Change-Id: I6b0770c5cd0db3e374f63f9c437fa9e8a77d3f8d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Martin Smith <martin.smith@qt.io>