qdoc: Introduce 'warninglimit' configuration variable
This variable is primarily meant for testing the documentation builds
in the CI system; if the warning count exceeds the set limit, QDoc
uses the total number of warnings as exit code.

The warning limit is enforced when either warninglimit.enabled = true,
or QDOC_ENABLE_WARNINGLIMIT environment variable is set.

This also provides a method to override the CI failing a documentation
build; the author can increase the warning limit to let the build
pass, hopefully with a good excuse in the code review.

Change-Id: If455a9636f2c63b41798523dd4acb3183a00229c
Reviewed-by: Martin Smith <martin.smith@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>