qdoc: Allow setting -no-link-errors via environment variable
Testing documentation builds in the CI will (at least initially)
ignore dependencies between doc modules, so link error count is not
a reliable metric to pass or reject documentation changes.

Allow CI to set an environment variable to disable linking errors,
as passing the -no-link-errors option to QDoc is not trivial in the
current build system.

Task-number: QTQAINFRA-1320
Change-Id: I8e6efaca4a0889d6f59ce781e9f05f3b836e95e9
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>
Reviewed-by: Martin Smith <martin.smith@qt.io>