qdoc: Set up output directory at Generator::initialize()
Although QDoc configures the (format-specific) output directory when
generating the docs, we need to have it already when building the node
tree. This is because a relative URL is passed to nodes read from index
files, and that URL is based on the output subdirectory name.

This fixes a linking issue introduced with commit 4c682533.

Change-Id: I6a943af8b28cfed60d70d476bd92f220ca7090f7
Reviewed-by: Martin Smith <martin.smith@qt.io>