qdoc: Remove unnecessary call of location().push()
The function pushes the file name onto the location stack so that
if an error occurs while processing the macro, the error message
can contain the correct file location for the macro in error. But
this call was pushing the location of the file containing all the
macro definitions. If there is an error processing the macro, it
should refer to the location where the macro was used, because
that's where the error is, not in the macro definition. So the
call to push() was wrong to begin with. Besides that, the push
was never popped, which led to a fatal qdoc error in macOS.

Change-Id: Ib2f74556f6a30a557e79d5c85cd12c8b3077b99a
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>