qdoc: Add optional since clause to \value command
This update adds the capability to include a since clause
in a \value command. Without the since clause, a \value
command might look like this:

\value QtInfoMsg A message generated by the qInfo() function.

The same command with a since clause would look like this:

\value [since 5.5] QtInfoMsg A message generated by the qInfo() function.

The since clause must be encosed in square brackets and must appear
before the value name.

Task-number: QTBUG-40782
Change-Id: If6cbb831c6f355eabcb0ddfdc284d2c9453f6245
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>