windeployqt: winrt: Do not list translation files that are not deployed
Visual Studio projects for winrt usage rely on a proper list of files
that are deployed. We concat translation files into a single
qt_<prefix>.qm file using lconvert, but also mention every "sub file"
(qtbase_<prefix>.qm etc) in json. That breaks deployment on winrt and
thus should not be done on that platform.

Change-Id: I3086b923ce4703ec159202cfe8df9447cdd51571
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>