linguist: use endonym for the translated language
technically speaking, this is just as wrong as not translating the
language, as we could be translating to a different language than the
current localization. however, translating every language name to every
other language is unrealistic, so this seems like a reasonable choice.

using endonyms required reshuffling the titles in the message editor,
as the grammar is otherwise broken. in fact, it may make sense to put
the language into quotes to detach it further from the surroundings.

Change-Id: Ie8f467dec73a2a0a13c2bbe5416bc1a869fb29cc
Reviewed-by: hjk <hjk@qt.io>