qdoc: Partial refactor of Node hierarchy
Ironically, The Node suclass CollectionNode should not be a
subclass of Aggregate because CollectionNode never has children.
It only has members, which are children of other Aggregate nodes.
This update makes CollectionNode a subclass of LeafNode instead
of Aggregate. This refactoring is necessary to implement the task
QTBUG-45725.

Change-Id: I60b2dcdaa3da1d138a0e6b4eb3d4f78cba0053d4
Task-number: QTBUG-45725
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>