qdoc: Fix retrieval of output directory
Config::getOutputDir() should not query the use of output subdirectory
from Generator; we need the output directory already when reading index
files, but its real value will be set later on in format-specific
initialization.

Instead, read it directly from config. This prevents an issue where
links are assigned a relative path to the subdirectory even when
'nosubdirs' config is used.

Change-Id: Ifc62336217b47ea3991f63ccc48248b3380c2b5c
Reviewed-by: Martin Smith <martin.smith@qt.io>