qdoc: Correct handling of functions declared in Q_OBJECT
These functions were being eliminated from the documentation in
the clang visitor, but this doesn't when the functions actually
have documentation. This update moves the test for whether they
should be documented or not to the Generator class.

This update also avoids reporting a qdoc warning for a typedef
defined in Q_GADGET.

Change-Id: Icca7297ec2123c684203d225d2bda96e50dacd8c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>