qdoc: Disable clang errors when the class is internal
From time to time, we have classes added that are marked both
\internal and \preliminary in their \class comment. This was a
race condition that set the status of the class documentation
to either internal or preliminary depending on the order of the
\internal and \preliminary commands. But \preliminary should
only be used when the class is meant to be included in the public
API marked as preliminary, so users will understand that it might
be changed before it is stable. So this update lets \internal win
when both commands appear in the \class comment, regardless of
their order.

This update also prevents clang parsing errors from being reported
for \fn commands where the class containing the member function is
marked \internal. This eliminates a lot of qdoc warnings.

Change-Id: Id1ecec4bdd573ae81fa5d589dfdd4afc4b313825
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>