macdeployqt: Implement selection of file system for .dmg file
With macOS High Sierra, Apple machines using SSD see their file
system changed for APFS automatically.

hdiutil by default uses the current file system to build the dmg file
which in this case makes them unusable on older version of macOS that
don't have support for APFS.

This patch adds an additional option to macdeployqt to select the
file system to use when building the .dmg. It defaults to HFS+ which
was the official file system used until now for all currently
supported version of macOS.

[ChangeLog][macdeployqt][feature] Added support for selecting the
file system type to use when building a .dmg file. Defaults to HFS+
to support a wider range of macOS versions.

Task-number: QTBUG-65844
Change-Id: Ic66856344f96c6536b224d13d309715b34eb0874
Reviewed-by: André Hartmann <aha_1980@gmx.de>