qdoc: Make setRelates() work for SharedCommentNode
When the \fn commands in a shared comment are global functions
that relate to some class or namespace, the shared comment node
and all the nodes sharing the comment must be marked as relating
to the class or namespace. This update makes the setRelates()
function virtual, and implements a special override in the
SharedCommentNode to mark all the nodes in the collective as
related to the class or namespace.

Also added hasSharedDoc() to return true if a node is
sharing a comment and the shared comment is not empty.
It might be the case that it can never return false if
the node is sharing a comment, but the function is
convenient anyway.

Change-Id: Ie064364e72fd150eacdd51a0131fe730311ee892
Reviewed-by: Martin Smith <martin.smith@qt.io>