qdoc: Fix handling of shared comment nodes in generate phase
Shared comments nodes were not detected and handled correctly in
certain cases. This update fixes those problems to avoid trying
to output documentation for the individual nodes, when they are
sharing a comment. Only the shared comment node should cause
documentation to be generated for the nodes sharing the comment.

Change-Id: I4ad7afc0964b1b6836dd5a140aa874784974b139
Reviewed-by: Martin Smith <martin.smith@qt.io>