qdoc: Remove test that skipped parsing \fn commands in prepare phase
It was thought unnecessary to parse the \fn commands during the prepare
phase, but it is required to determine whether the functions should be
marked internal and private. Without parsing the \fn commands during
the prepare phase, the documentation can't be assigned to the function,
so it looks like it is undocumented and is marked internal and private
in the index file. That was wrong, so now qdoc calls clang to parse the
\fn commands, even during the prepare phase. Unfortunately, this slows
qdoc by a significant amount.

Change-Id: I51ab621bb9d4ffff0a245d3a842d6e7b4558ca84
Reviewed-by: Martin Smith <martin.smith@qt.io>